Harden ci security

+++ pack.sh
@@ -1,6 +1,10 @@
 #!/bin/sh
 set -e
 BASE_NAME="$(jq '.id' ccmod.json | sed 's/^"//;s/"$//')"
+if [ -z "$BASE_NAME" ]; then
+	echo "Error: id not found in ccmod.json"
+	exit 1
+fi
 NAME="${BASE_NAME}-$(jq '.version' ccmod.json | sed 's/^"//;s/"$//').ccmod"
 rm -rf "$BASE_NAME"*
 pnpm install
@@ -12,7 +16,7 @@ cp -rf icon LICENSE plugin.js ./pack
 
 cd ./pack
 for file in $(find . -iname '*.json') $(find . -iname '*.json.patch') $(find . -iname '*.json.patch.cond'); do
-    jq '.' ../$file -c > $file
+	jq '.' ../$file -c >$file
 done
 cp ../ccmod.json .
 rm -rf icon/icon240.png
